implements fd_pread in wasi and adds zig wasi example
This implements fd_pread which is required when you don't buffer reads
in zig (particularly when your buffer isn't >= file length).

Advice on zig code welcome as it is a new language to me.

Note: The approach here is same as we use internall for GOOS=js as that
has an fdPread syscall. Just, we never implemented it for wasi, yet.

Signed-off-by: Adrian Cole <adrian@tetrate.io>