Adds logging.NewHostLoggingListenerFactory
This formalizes something I hand edit constantly, which is a listener
that only shows logging when it is a host function or called by the
host. This is important as some compilers create a large amount of
overhead functions, which can make thousands of lines around a host call
you are looking for.

Signed-off-by: Adrian Cole <adrian@tetrate.io>