wasi: benchmarks fd_readdir for both real and fake file systems (#910)
It is likely we'll have to special-case real files in wasi, as wasi-libc
makes behavioral choices based on presence of non-zero inode data. This
adds a defensive benchmark, so that we know what our performance base
case is. This adds both real and fake file systems as both patterns are
in use today.

Signed-off-by: Adrian Cole <adrian@tetrate.io>