cli: allows use of GOOS=js (#955)
This changes the CLI to detect if GOOS=js is likely in use and switches
accordingly. This makes it eaiser to find bugs and may eventually lead
to go being able to swap out node.js for wazero.

Note: GOOS=js a.k.a. gojs isn't complete, yet, so this helps identify
that as well. For example, similar to WASI there's no suppor for mkdir
yet.

```bash
$ GOOS=js GOARCH=wasm go test -o os.wasm -c os
$ wazero run os.wasm -test.v
```

Signed-off-by: Adrian Cole <adrian@tetrate.io>