Adds checks into the Makefile for 32-bit platforms
Developers don't commonly think about 32-bit platforms and in one
instance we broke a release due to this (quickly fixed in pre.7). This
ensures arm and 386, used by dapr and trivy respectively are tested. It
also helps prove these are in use to avoid a later developer removing
support by thinking they aren't.

Signed-off-by: Adrian Cole <adrian@tetrate.io>