Moves FreeBSD to compilation test only.
FreeBSD is not natively supported by GitHub actions or Docker. To test
it, we had to run vagrant which is both time-consuming (~7minutes), and
frequently fails to setup, resulting in false negatives. Meanwhile the
code related to FreeBSD is quite limited and doesn't change.

This changes our approach to testing FreeBSD to the same as how we deal
with 386, which is to do compilation tests only.

Signed-off-by: Adrian Cole <adrian@tetrate.io>