cli: adds an example Dockerfile
It is important for understanding the recent trends in Docker and
WebAssembly to see what a Dockerfile looks like in wazero. For example,
things that natively support wasm right now can take off the 4.5MB our
binary would cost, as well hide some of the configuration you can see,
such as coercing arguments or ENV. Since wasm isn't quite prime time in
Docker, some will be happy to use wazero instead or until it is ready.

Signed-off-by: Adrian Cole <adrian@tetrate.io>