data segment: encode the right prefix if a const expr is missing
If `OffsetExpression` is nil, we assume passive mode and use `0x1`

0x2 is only for the cases where the memory index is >=0,
i.e. even though 0 is a valid value, it only makes sense to use
when it is > 0. However, currently the spec does not support indices > 0.

So, for the time being, we only do the simple thing and correctly encode
passive segments as 0x1, otherwise defaulting to 0x0

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>