wasi: stop suggesting lookup of files by absolute path
WASI are recently suggesting by test that users who provide an
absolute path should fail via EPERM. Not all compilers deal with
pre-opens the same way. Notably, zig is pass-through, so it doesn't
expect to wash leading slashes around this rule.

While the jury is out on if we should break users as wasi-testsuite
wants us to, or not.. Meanwhile, we can stop suggesting use of absolute
paths, as relative ones work ("/" is the implied workdir in WASI).

See #1266

Signed-off-by: Adrian Cole <adrian@tetrate.io>