wasm: type ModuleInstance.Closed as atomic.Uint64
atomic.CompareAndSwapUint64 requires 64 bit alignment on the uint64
pointer passed to it. This is not easy to guarantee across platforms and
may cause invalid pointer faults. However, atomic.Unit64 is guaranteed
to have correct alignment.

See https://pkg.go.dev/sync/atomic#pkg-note-BUG for the quirky behaviors
in atomic.

This change does alter the expected access behavior to the Closed field,
but it appears that it's not (yet) consumed.

Signed-off-by: Tristan Willy <tristan.willy@gmail.com>