lifts emscripten.InstantiateForModule out of experimental
We added emscripten.InstantiateForModule into an experimental package so
that we could make a patch release and include it. Now that we've
changed api.Function, our next release needs to be a minor. This lifts
InstantiateForModule into the normal emscripten package as we know it
works and there's no known better signature to use for it.

Signed-off-by: Adrian Cole <adrian@tetrate.io>