Adds AccessMode Write and Pwrite to platform.File
This adds `file.AccessMode` to substitute for type checking that
formerly was used to see if the file might be writeable (e.g. casting to
`io.Writer` so that you can check before actually writing).

This allows us to implement `File.Write` and `File.Pwrite` which might
return `syscall.ENOSYS` on read-only implementations.

See #1417

Signed-off-by: Adrian Cole <adrian@tetrate.io>