Adds Read and Pread to platform.File
This adds `Read` and `Pread` to `platform.File` backfilling some tests
and porting others.

Notably, this no longer supports special casing `io.EOF` when the read
buffer is exactly the length of bytes remaining in the file, obviating
a follow-up read of length zero. The reason we can't do this internally
is `io.EOF` is a special flow control error Go defines, not an actual
syscall error. However, it is quite normal to not be able to handle
this and the calling code likely would loop around another time for the
same reason. Meanwhile the code is a lot less complex as we don't have
to test the io.EOF edge case.

Another note is I left stdin alone as I know @achille-roussel and maybe
@evacchi are working on non-blocking support there. I didn't want to
distract that or cause drift. We can address it later, and meanwhile, I
put in a TODO.

See #1417

Signed-off-by: Adrian Cole <adrian@tetrate.io>