Separates fs.File from os.File implementations of platform.File
Before, we had shared code which constantly checked if a file had a
descriptor or not. This was because the same type was used for files
backed by a real file (`os.File`) vs a potentially fake one (`fs.File`).
This untangles that by sifting logic into two types `fsFile` and
`osFile`. In doing so, we can eliminate more complexity around
windows-specific handling as it is a variation on `osFile`.

This also completes moving file-specific logic from `sys.FileEntry` into
`platform.File`, such as the responsibilty to cache file properties.
* push how to implement `rewinddir` into the `platform.File` type as
  `Seek(0,SeekStart)`. This removes the need to expose `File.Path` which
  is often incorrect. `fs.File` implementations receive a `name` on
  `fs.Open` which isn't usually an absolute path.
* push how to set append mode into the `platform.File` type as
  `IsAppend` and `SetAppend`. This removes the other case where
  re-opening a file happened externally.
* push cached inode into the file as `Ino` on `plaform.File`.

Signed-off-by: Adrian Cole <adrian@tetrate.io>