Removes AccessMode from platform.File
In the beginning, we checked by type if a command like read or write
might succeed. Temporarily, we switched to exposing the AccessMode,
which could be incorrect or drift. Now, we use the underlying error
instead. Removing `platform.AccessMode` was a TODO left until the code
was consolidated enough to address it coherently.

Signed-off-by: Adrian Cole <adrian@tetrate.io>