fd_readdir: reverts to special cased dirent cache
This reverts to a design similar to, but more performant than 1.2.0+.
Due to missing benchmarks, we accidentally regressed performance while
attempting to clean the internal implementation of directory pagination.
For reasons described in #1529, the refactoring was going the wrong way
anyway. This goes back to the better performing impl we had before,
makes it perform even better, and backfills missing documentation.

Signed-off-by: Adrian Cole <adrian@tetrate.io>