readdir: defer re-opening when Seek(0,SeekStart) on a directory
This defers re-opening the directory on `Seek(0, SeekStart)`.

I noticed that in the current code, we unconditionally seek to zero
before reading a directory. We already re-open the directory on the
first call to `Readdir` on windows. Moreover, the errno in `Seek` is
mostly about the position/location. Zero is always valid, so we don't
need to return an error.

If we defer instead we get multiple benefits:
* we don't re-open a windows directory twice on `Readdir` after
  `Seek(0, SeekStart)`
* we also don't re-open a fs.File backed directory, if `Seek(0,
  SeekStart)` was called twice.

I also re-jigged the tests a bit to trigger this even if testing that
this prevents called twice is a bit too tricky to attempt right now.

Signed-off-by: Adrian Cole <adrian@tetrate.io>