Allows wrapped filesystems to return dirent.ino except in windows
This allows `fs.File` backed implementations to return inode data the
same way regardless of it returns an `os.File` or a `fs.File`
implementing `Readdir`. Notably, this will not work in windows.

This proves implementations are the same by adding tests with a wrapper
implementing functions we use from os.File.

This might close #1553 depending on if @GeorgeMac is ok with the
limitations. This would not help with virtual files as the `fs.FileInfo`
is still special cased to platform-specific data.

Signed-off-by: Adrian Cole <adrian@tetrate.io>