fsapi: migrates PollRead to Poll with Pflag
This migrates PollRead to Poll so that it can support read, write or a
combination via flags. This should allow `poll_oneoff` to be implemented
even if internally it may need to use a deadline decrementing approach
to cycle through each file instead of via multiple at the same time.

The benefit of doing things this way is we don't need to top-level a
concept of file descriptor, which complicates the design of `fsapi.FS`
which is far easier to implement if there are no "multi-file" apis on
it.

This intentionally doesn't implement the approach at the same time, as
the goal is to create the api to implement. Doing so also puts a lot
less pressure on the task to complete migration off syscall types.

Signed-off-by: Adrian Cole <adrian@tetrate.io>