sysfs: decouples FS and File from the syscall package
This decouples FS and File from the syscall package by passing through
atim and mtim as numbers instead of structs. This makes it easier to
integrate with `os.Chtimes` as well possible to compile with plan9.

In doing so, this unsupports the `UTIME_NOW` constant, which was
problematic anyway. wasip1 defines `UTIME_NOW` as the value of the
real-time clock, which isn't necessarily the same as the filesystem
clock. By unsupporting `UTIME_NOW` we instead read from the correct
clock.

Signed-off-by: Adrian Cole <adrian@tetrate.io>