Exposes writeable filesystem as experimentalsys.FS
This exposes our internal filesystem as `experimentalsys.FS`,
configured with `sysfs.FSConfig`. This includes all the features used in
the CLI including the following:

* `sysfs.NewDirFS` - like `os.DirFS`
* `sysfs.NewReadFS` - masks write commands
* `sysfs.Adapt` - adapts an `fs.FS`

Implementors can use one of these or decorate by embedding one. When
writing from scratch, embed `UnimplementedFS/File` accordingly.

Here's an example of how to use this until it is no longer experimental:
```
import (
	"github.com/tetratelabs/wazero"
	"github.com/tetratelabs/wazero/experimental/sysfs"
)

root := sysfs.NewDirFS(".")

moduleConfig = wazero.NewModuleConfig().
	WithFSConfig(wazero.NewFSConfig().(sysfs.FSConfig).WithSysFSMount(root, "/"))
```

Note: This does not yet expose a general testing framework, but someone
can try to do that in another PR. It is tricky for reasons including
windows portability which become better when we no longer have to
support Go 1.18 (when Go 1.21 is released).

Signed-off-by: Adrian Cole <adrian@tetrate.io>