sysfs: Makes ReadFS and AdaptFS embeddable
This exports the experimental `sysfs.ReadFS` and `sysfs.AdaptFS` types,
to developers can choose to embed them. This makes certain decoration
use cases easier.

Note: We can't do the same without a risk of drift in `sysfs.DirFS`, but
that isn't likely a problem as people are used to `os.DirFS` which also
doesn't export its type.

Signed-off-by: Adrian Cole <adrian@tetrate.io>